Made location info use size_t instead of int for indexes.

+++ include/libscc/parser.h
@@ -213,10 +213,10 @@ typedef union YYSTYPE YYSTYPE;
 
 #define YYLTYPE YYLTYPE
 typedef struct YYLTYPE {
-    int first_line;
-    int first_column;
-    int last_line;
-    int last_column;
+    size_t first_line;
+    size_t first_column;
+    size_t last_line;
+    size_t last_column;
     const char * filename;
 } YYLTYPE;
 #define YYLTYPE_IS_DECLARED 1
